requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
zxcvbn==4.4.28
python-multipart==0.0.9
pydantic==1.10.15
yt-dlp==2024.3.10
aiofiles==23.2.1
